requirements.txt
ebooklib
requests
pydub
SpeechRecognition
mutagen
PyYAML
pytest
pytest-cov
tqdm
fastapi
uvicorn[standard]
websockets
python-multipart
aiofiles
